+++ handle_octal.c
@@ -7,7 +7,7 @@
  * @buf_index: The current index in the buffer
  * @count: The character count
  */
-void handle_octal(va_list args, char *buf, int *buf_index, int *count)
+void handle_octal(va_list args, char *buf, size_t *buf_index, int *count)
 {
 	unsigned int num = va_arg(args, unsigned int);
 	char octal_str[12];
